LPS-62701 Move interface out of base
