+++ modules/frontend/frontend-taglib/src/main/java/com/liferay/frontend/taglib/servlet/taglib/ManagementBarButtonsTag.java
@@ -15,7 +15,6 @@
 package com.liferay.frontend.taglib.servlet.taglib;
 
 import com.liferay.frontend.taglib.servlet.ServletContextUtil;
-import com.liferay.frontend.taglib.servlet.taglib.base.BarTag;
 import com.liferay.taglib.util.IncludeTag;
 
 import javax.servlet.jsp.JspException;
