+++ modules/frontend/frontend-taglib/src/main/java/com/liferay/frontend/taglib/servlet/taglib/base/BaseBarTag.java
@@ -15,6 +15,7 @@
 package com.liferay.frontend.taglib.servlet.taglib.base;
 
 import com.liferay.frontend.taglib.servlet.ServletContextUtil;
+import com.liferay.frontend.taglib.servlet.taglib.BarTag;
 import com.liferay.taglib.util.IncludeTag;
 
 import javax.servlet.jsp.PageContext;
